+++ package_monitor/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "package_monitor.apps.PackageMonitorConfig"
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 __title__ = "Package Monitor"